failure for school data
I had a very hard time in converting the character values into numeric values.
For the percent values, I had 100 and 18.3
If I format it with 4.1, it would read 100 as 10
If I format it with 5 it wouldn't be able to read 18.3

To fix it I need bucn of if statement to find whether there is a '.' in the character value.